fix: add caching for fetching badges
