+++ src/graphql/getters/getSeasonBadges/index.ts
@@ -1,2 +1,4 @@
 export * from './getSeasonBadges';
+export * from './mappers';
+export * from './types';
 export * from './helpers';
